--- .env
@@ -1,3 +1,0 @@
-VITE_BACKEND_URL= https://api.bmsamay.com
-VITE_BACKEND_URL_2 = https://api2.bmsamay.com
-VITE_YOUTUBE_API_KEY=
