+++ package.json
@@ -19,7 +19,6 @@
     "@radix-ui/react-slider": "^1.2.0",
     "@radix-ui/react-slot": "^1.1.0",
     "@radix-ui/react-tabs": "^1.1.0",
-    "@types/jwt-decode": "^3.1.0",
     "axios": "^1.7.2",
     "chess.js": "1.0.0-beta.8",
     "class-variance-authority": "^0.6.1",
@@ -54,7 +53,6 @@
     "@types/node": "^20.8.0",
     "@types/react": "^18.0.37",
     "@types/react-dom": "^18.0.11",
-    "@types/react-icons": "^3.0.0",
     "@typescript-eslint/eslint-plugin": "^5.59.0",
     "@typescript-eslint/parser": "^5.59.0",
     "@vitejs/plugin-react": "^4.0.0",
