+++ src/main.tsx
@@ -3,23 +3,23 @@ import ReactDOM from 'react-dom/client';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import { HelmetProvider } from 'react-helmet-async';
 import App from './App.tsx';
-import { LandingPage } from './pages/landing.tsx';
-import Signup from './pages/signup.tsx';
+import { LandingPage } from './pages/Landing.tsx';
+import Signup from './pages/SignUp.tsx';
 import '@/styles/globals.css';
-import SignUpCallback from './pages/signupcallback.tsx';
-import Welcome from './pages/welcome.tsx';
-import Community from './pages/community.tsx';
-import Blunder from './pages/somethingwentwrong.tsx';
+import SignUpCallback from './pages/SignUpCallback.tsx';
+import Welcome from './pages/Welcome.tsx';
+import Community from './pages/Community.tsx';
+import Blunder from './pages/SomethingWentWrong.tsx';
 import ErrorBoundary from './components/ErrorBoundary.tsx';
-import PrivacyPolicy from './pages/privacypolicy.tsx';
+import PrivacyPolicy from './pages/PrivacyPolicy.tsx';
 import ComingSoon from './components/comingsoon.tsx';
 import ChessNews from './pages/ChessNews.tsx';
 import ChessTutorials from './pages/ChessTutorials.tsx';
-import Chat from './pages/chat.tsx';
-import DiamondGift from './pages/gift.tsx';
+import Chat from './pages/Chat.tsx';
+import DiamondGift from './pages/Gift.tsx';
 import CookieConsentWrapper from './components/CookieConsentWrapper.tsx';
 import PageNotFound from './pages/404NotFound.tsx';
-import GuessTheElo from './pages/gte.tsx';
+import GuessTheElo from './pages/GTE.tsx';
 import Learn from './pages/Learn.tsx';
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
