+++ src/pages/GTE.tsx
